made a lot of changes; NOTE: COMMIT  WILL NOT COMPILE! - will remove this commit in the coming week and finish up v0.2.1

+++ src/tree/search/RnnModule.f90
@@ -2,9 +2,9 @@ submodule(KdTree:SearchSubmod) RnnModule
     implicit none 
     contains 
         module procedure rNN
-            real(kind=real64)                           :: delta
-            type(nodePtr), allocatable                  :: tmp(:)
-            integer                                     :: axis
+            real(kind=real64)           :: delta
+            type(nodePtr), allocatable  :: tmp(:)
+            integer(int64)              :: axis
             
             if (associated(curr)) then
